+++ ntex/src/web/extract.rs
@@ -1,9 +1,9 @@
 //! Request extractors
-use std::{future::Future, pin::Pin, task::Context, task::Poll};
+use std::future::Future;
 
 use super::error::ErrorRenderer;
 use super::httprequest::HttpRequest;
-use crate::{http::Payload, util::BoxFuture, util::Ready};
+use crate::http::Payload;
 
 /// Trait implemented by types that can be extracted from request.
 ///
@@ -12,11 +12,11 @@ pub trait FromRequest<Err>: Sized {
     /// The associated error which can be returned.
     type Error;
 
-    /// Future that resolves to a Self
-    type Future: Future<Output = Result<Self, Self::Error>>;
-
     /// Convert request to a Self
-    fn from_request(req: &HttpRequest, payload: &mut Payload) -> Self::Future;
+    fn from_request(
+        req: &HttpRequest,
+        payload: &mut Payload,
+    ) -> impl Future<Output = Result<Self, Self::Error>>;
 }
 
 /// Optionally extract a field from the request
@@ -26,7 +26,7 @@ pub trait FromRequest<Err>: Sized {
 /// ## Example
 ///
 /// ```rust
-/// use ntex::{http, util::Ready};
+/// use ntex::http;
 /// use ntex::web::{self, error, App, HttpRequest, FromRequest, DefaultError};
 /// use rand;
 ///
@@ -37,13 +37,12 @@ pub trait FromRequest<Err>: Sized {
 ///
 /// impl<Err> FromRequest<Err> for Thing {
 ///     type Error = error::Error;
-///     type Future = Ready<Self, Self::Error>;
 ///
-///     fn from_request(req: &HttpRequest, payload: &mut http::Payload) -> Self::Future {
+///     async fn from_request(req: &HttpRequest, payload: &mut http::Payload) -> Result<Self, Self::Error> {
 ///         if rand::random() {
-///             Ready::Ok(Thing { name: "thingy".into() })
+///             Ok(Thing { name: "thingy".into() })
 ///         } else {
-///             Ready::Err(error::ErrorBadRequest("no luck").into())
+///             Err(error::ErrorBadRequest("no luck").into())
 ///         }
 ///     }
 /// }
@@ -66,25 +65,24 @@ pub trait FromRequest<Err>: Sized {
 /// ```
 impl<T, Err> FromRequest<Err> for Option<T>
 where
-    T: FromRequest<Err> + 'static,
+    T: FromRequest<Err>,
     Err: ErrorRenderer,
     <T as FromRequest<Err>>::Error: Into<Err::Container>,
 {
     type Error = Err::Container;
-    type Future = BoxFuture<'static, Result<Option<T>, Self::Error>>;
 
     #[inline]
-    fn from_request(req: &HttpRequest, payload: &mut Payload) -> Self::Future {
-        let fut = T::from_request(req, payload);
-        Box::pin(async move {
-            match fut.await {
-                Ok(v) => Ok(Some(v)),
-                Err(e) => {
-                    log::debug!("Error for Option<T> extractor: {}", e.into());
-                    Ok(None)
-                }
+    async fn from_request(
+        req: &HttpRequest,
+        payload: &mut Payload,
+    ) -> Result<Option<T>, Self::Error> {
+        match T::from_request(req, payload).await {
+            Ok(v) => Ok(Some(v)),
+            Err(e) => {
+                log::debug!("Error for Option<T> extractor: {}", e.into());
+                Ok(None)
             }
-        })
+        }
     }
 }
 
@@ -95,7 +93,7 @@ where
 /// ## Example
 ///
 /// ```rust
-/// use ntex::{http, util::Ready};
+/// use ntex::http;
 /// use ntex::web::{self, error, App, HttpRequest, FromRequest};
 /// use rand;
 ///
@@ -106,13 +104,12 @@ where
 ///
 /// impl<Err> FromRequest<Err> for Thing {
 ///     type Error = error::Error;
-///     type Future = Ready<Thing, Self::Error>;
 ///
-///     fn from_request(req: &HttpRequest, payload: &mut http::Payload) -> Self::Future {
+///     async fn from_request(req: &HttpRequest, payload: &mut http::Payload) -> Result<Thing, Self::Error> {
 ///         if rand::random() {
-///             Ready::Ok(Thing { name: "thingy".into() })
+///             Ok(Thing { name: "thingy".into() })
 ///         } else {
-///             Ready::Err(error::ErrorBadRequest("no luck").into())
+///             Err(error::ErrorBadRequest("no luck").into())
 ///         }
 ///     }
 /// }
@@ -133,38 +130,34 @@ where
 /// ```
 impl<T, E> FromRequest<E> for Result<T, T::Error>
 where
-    T: FromRequest<E> + 'static,
-    T::Error: 'static,
+    T: FromRequest<E>,
     E: ErrorRenderer,
 {
     type Error = T::Error;
-    type Future = BoxFuture<'static, Result<Result<T, T::Error>, Self::Error>>;
 
     #[inline]
-    fn from_request(req: &HttpRequest, payload: &mut Payload) -> Self::Future {
-        let fut = T::from_request(req, payload);
-        Box::pin(async move {
-            match fut.await {
-                Ok(v) => Ok(Ok(v)),
-                Err(e) => Ok(Err(e)),
-            }
-        })
+    async fn from_request(
+        req: &HttpRequest,
+        payload: &mut Payload,
+    ) -> Result<Self, Self::Error> {
+        match T::from_request(req, payload).await {
+            Ok(v) => Ok(Ok(v)),
+            Err(e) => Ok(Err(e)),
+        }
     }
 }
 
 #[doc(hidden)]
 impl<E: ErrorRenderer> FromRequest<E> for () {
     type Error = E::Container;
-    type Future = Ready<(), E::Container>;
 
     #[inline]
-    fn from_request(_: &HttpRequest, _: &mut Payload) -> Self::Future {
-        Ok(()).into()
+    async fn from_request(_: &HttpRequest, _: &mut Payload) -> Result<(), E::Container> {
+        Ok(())
     }
 }
 
 macro_rules! tuple_from_req ({$fut_type:ident, $(($n:tt, $T:ident)),+} => {
-
     /// FromRequest implementation for a tuple
     #[allow(unused_parens)]
     impl<Err: ErrorRenderer, $($T: FromRequest<Err> + 'static),+> FromRequest<Err> for ($($T,)+)
@@ -172,54 +165,11 @@ macro_rules! tuple_from_req ({$fut_type:ident, $(($n:tt, $T:ident)),+} => {
         $(<$T as $crate::web::FromRequest<Err>>::Error: Into<Err::Container>),+
     {
         type Error = Err::Container;
-        type Future = $fut_type<Err, $($T),+>;
 
-        fn from_request(req: &HttpRequest, payload: &mut Payload) -> Self::Future {
-            $fut_type {
-                items: <($(Option<$T>,)+)>::default(),
-                $($T: $T::from_request(req, payload),)+
-            }
-        }
-    }
-
-    pin_project_lite::pin_project! {
-        #[doc(hidden)]
-    pub struct $fut_type<Err: ErrorRenderer, $($T: FromRequest<Err>),+>
-    {
-        items: ($(Option<$T>,)+),
-        $(#[pin] $T: $T::Future),+
-    }
-    }
-
-    impl<Err: ErrorRenderer, $($T: FromRequest<Err>),+> Future for $fut_type<Err, $($T),+>
-    where
-        $(<$T as $crate::web::FromRequest<Err>>::Error: Into<Err::Container>),+
-    {
-        type Output = Result<($($T,)+), Err::Container>;
-
-        fn poll(self: Pin<&mut Self>, cx: &mut Context<'_>) -> Poll<Self::Output> {
-            let this = self.project();
-
-            let mut ready = true;
-            $(
-                if this.items.$n.is_none() {
-                    match this.$T.poll(cx) {
-                        Poll::Ready(Ok(item)) => {
-                            this.items.$n = Some(item);
-                        }
-                        Poll::Pending => ready = false,
-                        Poll::Ready(Err(e)) => return Poll::Ready(Err(e.into())),
-                    }
-                }
-            )+
-
-                if ready {
-                    Poll::Ready(Ok(
-                        ($(this.items.$n.take().unwrap(),)+)
-                    ))
-                } else {
-                    Poll::Pending
-                }
+        async fn from_request(req: &HttpRequest, payload: &mut Payload) -> Result<($($T,)+), Err::Container> {
+            Ok((
+                $($T::from_request(req, payload).await.map_err(|e| e.into())?,)+
+            ))
         }
     }
 });