+++ ntex/src/web/types/json.rs
@@ -163,30 +163,28 @@ where
     T: DeserializeOwned + 'static,
 {
     type Error = JsonPayloadError;
-    type Future = BoxFuture<'static, Result<Self, Self::Error>>;
 
-    #[inline]
-    fn from_request(req: &HttpRequest, payload: &mut Payload) -> Self::Future {
+    async fn from_request(
+        req: &HttpRequest,
+        payload: &mut Payload,
+    ) -> Result<Self, Self::Error> {
         let req2 = req.clone();
         let (limit, ctype) = req
             .app_state::<JsonConfig>()
             .map(|c| (c.limit, c.content_type.clone()))
             .unwrap_or((32768, None));
 
-        let fut = JsonBody::new(req, payload, ctype).limit(limit);
-        Box::pin(async move {
-            match fut.await {
-                Err(e) => {
-                    log::debug!(
-                        "Failed to deserialize Json from payload. \
-                         Request path: {}",
-                        req2.path()
-                    );
-                    Err(e)
-                }
-                Ok(data) => Ok(Json(data)),
+        match JsonBody::new(req, payload, ctype).limit(limit).await {
+            Err(e) => {
+                log::debug!(
+                    "Failed to deserialize Json from payload. \
+                     Request path: {}",
+                    req2.path()
+                );
+                Err(e)
             }
-        })
+            Ok(data) => Ok(Json(data)),
+        }
     }
 }